refs and fixes
